Kdv Tutarı hesaplayan program yapıldı.

+++ ProgramThatCalculatesVATamount/src/main/java/org/example/Main.java
@@ -10,7 +10,7 @@ public class Main {
         System.out.println("Lutfen Para degerini giriniz: ");
 
         double kdvsizFiyat= sc.nextInt();
-        double kdvliFiyat=(double)(((kdvsizFiyat*18)/100)+kdvsizFiyat);
+        double kdvliFiyat=(((kdvsizFiyat*18)/100)+kdvsizFiyat);
         double kdvTutari=(kdvliFiyat-kdvsizFiyat);
 
         System.out.println("KDV'siz fiyat = "+kdvsizFiyat+"KDV'li fiyat = "+kdvliFiyat+"KDVTutari = "+kdvTutari);
